+++ Code/EasyMusicPlayer/CurrentTrackInformation.swift
@@ -1,0 +1,8 @@
+import MediaPlayer
+
+struct CurrentTrackInformation {
+    let track: MPMediaItem?
+    let index: Int
+    var number: Int { index + 1 }
+    var duration: TimeInterval { track?.playbackDuration ?? 0 }
+}