+++ Code/EasyMusicPlayer/MusicPlayerInformation.swift
@@ -7,10 +7,3 @@ struct MusicPlayerInformation {
     let isPlaying: Bool
     let repeatMode: RepeatMode
 }
-
-struct CurrentTrackInformation {
-    let item: MPMediaItem?
-    let index: Int
-    var number: Int { index + 1 }
-    var duration: TimeInterval { item?.playbackDuration ?? 0 }
-}