+++ Code/EasyMusicPlayer/MusicPlayerState.swift
@@ -4,6 +4,7 @@ enum MusicPlayerState {
     case play
     case pause
     case stop
+    case reset
     case repeatMode(RepeatMode)
     case clock(TimeInterval)
     case error(MusicPlayerError)